# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vehicle Management API",
        "description": "A comprehensive API for managing vehicles, documents, maintenance records, and vehicle exchanges with role-based access control.",
        "contact": {
            "name": "API Support",
            "email": "support@vehiclemanagement.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/vehicles/{vehicle_id}/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get vehicle documents",
                "description": "Retrieve all documents for a specific vehicle",
                "operationId": "5e04a7ae998e2e162f5561c6d7d609cc",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Documents retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/VehicleDocument"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Vehicle not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload vehicle document",
                "description": "Upload a new document for a vehicle",
                "operationId": "e3a4b2145cb673457af63a93cd1c0513",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "type",
                                    "expiration_date"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "carte_grise",
                                            "assurance",
                                            "controle_technique"
                                        ],
                                        "example": "assurance"
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    },
                                    "file": {
                                        "description": "Document file (PDF, JPG, JPEG, PNG, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Document created successfully"
                                        },
                                        "document": {
                                            "$ref": "#/components/schemas/VehicleDocument"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Vehicle not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicles/{vehicle_id}/documents/{document_id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get document details",
                "description": "Retrieve details of a specific vehicle document",
                "operationId": "8f5d7f2b1392bc8434fa6703abf4db2d",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "document_id",
                        "in": "path",
                        "description": "Document ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document details retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "document": {
                                            "$ref": "#/components/schemas/VehicleDocument"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Document not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicles/{vehicle_id}/documents/expiring": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get expiring documents",
                "description": "Get documents that are expiring soon for a specific vehicle",
                "operationId": "48facaffbaf0b427a78fe70d2df2e550",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Number of days to check for expiring documents",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expiring documents retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "documents": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/VehicleDocument"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Vehicle not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "VehicleDocument": {
                "title": "Vehicle Document",
                "description": "Vehicle document model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "vehicle_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "carte_grise",
                            "assurance",
                            "controle_technique"
                        ],
                        "example": "assurance"
                    },
                    "expiration_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    },
                    "file_path": {
                        "type": "string",
                        "example": "documents/doc_2024-01-01_abc123.pdf",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}